# Performance optimization for SvgPointCollectionConverter parser