--- a/Source/DataTypes/SvgPointCollection.cs
+++ b/Source/DataTypes/SvgPointCollection.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Globalization;
 using System.Text;
+using Svg.Helpers;
 
 namespace Svg
 {
@@ -44,8 +45,10 @@ namespace Svg
     /// <summary>
     /// A class to convert string into <see cref="SvgPointCollection"/> instances.
     /// </summary>
-    internal class SvgPointCollectionConverter : TypeConverter
+    public class SvgPointCollectionConverter : TypeConverter
     {
+        private static readonly char[] SplitChars = new[] { ' ', '\t', '\n', '\r', ',' };
+
         /// <summary>
         /// Converts the given object to the type of this converter, using the specified context and culture information.
         /// </summary>
@@ -60,18 +63,37 @@ namespace Svg
         {
             if (value is string s)
             {
-                var coords = s.AsSpan().Trim();
-                var state = new CoordinateParserState(ref coords);
-                var result = new SvgPointCollection();
-                while (CoordinateParser.TryGetFloat(out var pointValue, ref coords, ref state))
-                {
-                    result.Add(new SvgUnit(SvgUnitType.User, pointValue));
-                }
-
-                return result;
+                return Parse(s.AsSpan());
             }
 
             return base.ConvertFrom(context, culture, value);
+        }
+
+        public static SvgPointCollection Parse(ReadOnlySpan<char> points)
+        {
+#if false
+            var coords = points.Trim();
+            var state = new CoordinateParserState(ref coords);
+            var result = new SvgPointCollection();
+            while (CoordinateParser.TryGetFloat(out var pointValue, ref coords, ref state))
+            {
+                result.Add(new SvgUnit(SvgUnitType.User, pointValue));
+            }
+            return result;
+#else
+            var collection = new SvgPointCollection();
+            var splitChars = SplitChars.AsSpan();
+            var parts = new StringSplitEnumerator(points, splitChars);
+
+            foreach (var part in parts)
+            {
+                var partValue = part.Value;
+                var result = StringParser.ToFloat(ref partValue);
+                collection.Add(new SvgUnit(SvgUnitType.User, result));
+            }
+
+            return collection;
+#endif
         }
     }
 }