@inforithmics 2022-03-15
I think this can now be merged because. The Roslyn Compiler can now do switches on ReadonlySpans
see issue.
https://github.com/dotnet/csharplang/issues/1881